AdminBase to set default values as options in class
